+++ Drop-In/src/main/java/com/braintreepayments/api/dropin/adapters/VaultedPaymentMethodsAdapter.java
@@ -1,5 +1,6 @@
 package com.braintreepayments.api.dropin.adapters;
 
+import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
@@ -7,25 +8,36 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import com.braintreepayments.api.PaymentMethod;
+import com.braintreepayments.api.dropin.DropInRequest;
 import com.braintreepayments.api.dropin.R;
 import com.braintreepayments.api.dropin.utils.PaymentMethodType;
 import com.braintreepayments.api.interfaces.PaymentMethodNonceCreatedListener;
 import com.braintreepayments.api.models.CardNonce;
+import com.braintreepayments.api.models.Configuration;
 import com.braintreepayments.api.models.PaymentMethodNonce;
 
 import java.util.List;
 
 import androidx.recyclerview.widget.RecyclerView;
 
+// NEXT MAJOR VERSION make this class package private
 public class VaultedPaymentMethodsAdapter extends RecyclerView.Adapter<VaultedPaymentMethodsAdapter.ViewHolder> {
 
-    private PaymentMethodNonceCreatedListener mSelectedListener;
-    private List<PaymentMethodNonce> mPaymentMethodNonces;
+    private final PaymentMethodNonceCreatedListener mSelectedListener;
+
+    private final List<PaymentMethodNonce> mPaymentMethodNonces;
+    private AvailablePaymentMethodNonceList mAvailablePaymentMethodNonces;
 
     public VaultedPaymentMethodsAdapter(PaymentMethodNonceCreatedListener listener,
                                         List<PaymentMethodNonce> paymentMethodNonces) {
         mSelectedListener = listener;
         mPaymentMethodNonces = paymentMethodNonces;
+    }
+
+    public void setup(Context context, Configuration configuration, DropInRequest dropInRequest, boolean googlePayEnabled, boolean unionpaySupported) {
+        mAvailablePaymentMethodNonces = new AvailablePaymentMethodNonceList(
+                context, configuration, mPaymentMethodNonces, dropInRequest, googlePayEnabled);
     }
 
     @Override
@@ -36,7 +48,7 @@ public class VaultedPaymentMethodsAdapter extends RecyclerView.Adapter<VaultedPa
 
     @Override
     public void onBindViewHolder(ViewHolder holder, int position) {
-        final PaymentMethodNonce paymentMethodNonce = mPaymentMethodNonces.get(position);
+        final PaymentMethodNonce paymentMethodNonce = mAvailablePaymentMethodNonces.get(position);
         PaymentMethodType paymentMethodType = PaymentMethodType.forType(paymentMethodNonce);
 
         holder.icon.setImageResource(paymentMethodType.getVaultedDrawable());
@@ -58,7 +70,11 @@ public class VaultedPaymentMethodsAdapter extends RecyclerView.Adapter<VaultedPa
 
     @Override
     public int getItemCount() {
-        return mPaymentMethodNonces.size();
+        return mAvailablePaymentMethodNonces.size();
+    }
+
+    public boolean hasCardNonce() {
+        return mAvailablePaymentMethodNonces.hasCardNonce();
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder {