+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree Android Drop-In Release Notes
 
+## unreleased
+
+* Fix issue where payment methods disabled on `DropInRequest` were displayed if previously vaulted (fixes #205)
+
 ## 5.0.1
 
 * Bump braintree_android version to 3.14.2 (fixes [#197](https://github.com/braintree/braintree-android-drop-in/issues/197))